requirements.txt
Pillow
torch
torchvision
rembg
tqdm
numpy
scikit-learn
onnxruntime